Exclude negative transients
If a negative transient is an asteroid, it corresponds to an asteroid
in the reference image. As reference images can be stacked, the
magnitude and position estimates can be unreliable. Moreover, the
observation date and time are unclear for a stacked image. As match2SSO
takes the observing time of the new image as the detection time of the
asteroid, this results in incorrect matches for negative transients.